+++ tests/testthat/test-biomes_get.R
@@ -1,4 +1,5 @@
 test_that("biomes_get returns a 31-layer SpatRaster", {
+  skip_if_no_raster()
   layers <- biomes_get()
   expect_s4_class(layers, "SpatRaster")
   expect_equal(terra::nlyr(layers), 31)